Moved the scarf specific code out of ProofObligations.py.  Moved the chc_report_project_scarf.py script to a new scarf_exporter directory.
